debt - move more window things to `window` layer
