+++ src/vs/platform/window/electron-sandbox/window.ts
